fastapi_app/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
transformers
torch
scikit-learn
streamlit

